+++ wheel/src/api.rs
@@ -5,10 +5,11 @@ use crate::adapt_response::adapt_response;
 use clvmr::allocator::Allocator;
 use clvmr::chia_dialect::ChiaDialect;
 use clvmr::cost::Cost;
+use clvmr::deserialize_tree::{deserialize_tree, CLVMTreeBoundary};
 use clvmr::reduction::Response;
 use clvmr::run_program::run_program;
 use clvmr::serde::{node_from_bytes, serialized_length_from_bytes};
-use clvmr::serialize::{parse_triples, ParsedTriple};
+use clvmr::deserialize_tree::{parse_triples, ParsedTriple};
 use clvmr::{LIMIT_HEAP, LIMIT_STACK, MEMPOOL_MODE, NO_UNKNOWN_OPS};
 use pyo3::prelude::*;
 use pyo3::types::PyTuple;
@@ -43,14 +44,14 @@ pub fn run_serialized_chia_program(
     adapt_response(py, allocator, r)
 }
 
-fn tuple_for_parsed_triple(py: Python<'_>, p: &ParsedTriple) -> PyObject {
+fn tuple_for_parsed_triple(py: Python<'_>, p: &CLVMTreeBoundary) -> PyObject {
     let tuple = match p {
-        ParsedTriple::Atom {
+        CLVMTreeBoundary::Atom {
             start,
             end,
             atom_offset,
         } => PyTuple::new(py, [*start, *end, *atom_offset as u64]),
-        ParsedTriple::Pair {
+        CLVMTreeBoundary::Pair {
             start,
             end,
             right_index,
@@ -60,9 +61,9 @@ fn tuple_for_parsed_triple(py: Python<'_>, p: &ParsedTriple) -> PyObject {
 }
 
 #[pyfunction]
-fn deserialize_as_triples(py: Python, blob: &[u8]) -> PyResult<Vec<PyObject>> {
+fn deserialize_as_tree(py: Python, blob: &[u8]) -> PyResult<Vec<PyObject>> {
     let mut cursor = io::Cursor::new(blob);
-    let r = parse_triples(&mut cursor)?;
+    let r = deserialize_tree(&mut cursor)?;
     let r = r.iter().map(|pt| tuple_for_parsed_triple(py, pt)).collect();
     Ok(r)
 }
